# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Spectroradiometer Analyzer API",
    "version": "1.0.0"
  },
  "paths": {
    "/api/v1/sensors/{sensor_id}/spectrum": {
      "get": {
        "tags": [
          "Hardware"
        ],
        "summary": "Get Spectrum",
        "operationId": "get_spectrum_api_v1_sensors__sensor_id__spectrum_get",
        "parameters": [
          {
            "name": "sensor_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Sensor Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpectralData"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sensors/analyze": {
      "post": {
        "tags": [
          "Hardware",
          "Análisis"
        ],
        "summary": "Analyze Spectra",
        "description": "Configura de forma atómica y adquiere los datos de los sensores solicitados.",
        "operationId": "analyze_spectra_api_v1_sensors_analyze_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalysisRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AnalysisRequest": {
        "properties": {
          "sensor_target": {
            "type": "string",
            "enum": [
              "MS-711",
              "MS-712",
              "Merge"
            ],
            "title": "Sensor Target",
            "description": "Sensor a leer, o 'Merge' para adquirir y fusionar ambos."
          },
          "exposure_time_ms": {
            "type": "integer",
            "maximum": 5000.0,
            "minimum": 10.0,
            "title": "Exposure Time Ms",
            "description": "Tiempo de exposición para la lectura (10-5000ms).",
            "default": 10
          }
        },
        "type": "object",
        "required": [
          "sensor_target"
        ],
        "title": "AnalysisRequest",
        "description": "Payload para iniciar una configuración y lectura atómica."
      },
      "AnalysisResult": {
        "properties": {
          "merged_spectrum": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SpectralData"
              },
              {
                "type": "null"
              }
            ],
            "description": "Espectro fusionado e interpolado a 1nm de resolución"
          },
          "par": {
            "type": "number",
            "title": "Par",
            "description": "Radiación Fotosintéticamente Activa (W/m²), integrada de 400 a 700nm"
          },
          "ppfd": {
            "type": "number",
            "title": "Ppfd",
            "description": "Densidad de Flujo de Fotones Fotosintéticos (µmol/m²/s)"
          },
          "illuminance": {
            "type": "number",
            "title": "Illuminance",
            "description": "Iluminancia fotópica (lux)"
          },
          "total_irradiance": {
            "type": "number",
            "title": "Total Irradiance",
            "description": "Irradiancia total integrada sobre el rango completo (W/m²)"
          }
        },
        "type": "object",
        "required": [
          "par",
          "ppfd",
          "illuminance",
          "total_irradiance"
        ],
        "title": "AnalysisResult",
        "description": "Resultado del análisis espectral combinado MS-711 + MS-712."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "SpectralData": {
        "properties": {
          "wavelengths": {
            "items": {
              "type": "number"
            },
            "type": "array",
            "title": "Wavelengths",
            "description": "Array de longitudes de onda (nm)"
          },
          "irradiance": {
            "items": {
              "type": "number"
            },
            "type": "array",
            "title": "Irradiance",
            "description": "Irradiancia espectral (W/m2/um)"
          }
        },
        "type": "object",
        "required": [
          "wavelengths",
          "irradiance"
        ],
        "title": "SpectralData"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}